Moved automation settings to AutoPlayTab
